Toon Shader Implemented
If you want to add toonshading to a mesh, create a new material, and in the dropdown select "Shadergraphs/Toon".

Leave the shadow ramp texture as the default, and simply set the desired colour or texture as the diffuse. Specular highlights can be adjusted by the smoothness value, where 0 is most reflective and 1 is least.